Rewrote file struct to use linked list
File structure has been rewritten to use linked list instead of
contiguous memory. This is to make certain operations such as splitting
a line or appending a line to the previous line.

Only includes the functions necessary to read a file at the moment.
Further work on the writing to the file isn't necessary until the
ncurses interface is rewritten